update functions short path

+++ src/api/node_data.java
@@ -47,11 +47,11 @@ public interface node_data {
 	 * which can be used be algorithms 
 	 * @return
 	 */
-	public double getTag();
+	public int getTag();
 	/** 
 	 * Allows setting the "tag" value for temporal marking an node - common
 	 * practice for marking by algorithms.
 	 * @param t - the new value of the tag
 	 */
-	public void setTag(double t);
+	public void setTag(int t);
 }
